# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "VoxFlow",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/tasks": {
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "Create Task",
        "operationId": "create_task_api_v1_tasks_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskCreateRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskCreateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "List Tasks",
        "operationId": "list_tasks_api_v1_tasks_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "model_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Model Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_TaskSummary_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tasks/{task_id}": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "Get Task",
        "operationId": "get_task_api_v1_tasks__task_id__get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tasks/{task_id}/result": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "Get Result",
        "operationId": "get_result_api_v1_tasks__task_id__result_get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InferenceResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tasks/{task_id}/events": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "Task Events",
        "operationId": "task_events_api_v1_tasks__task_id__events_get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tasks/{task_id}/cancel": {
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "Cancel Task",
        "operationId": "cancel_task_api_v1_tasks__task_id__cancel_post",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Cancel Task Api V1 Tasks  Task Id  Cancel Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tasks/{task_id}/retry": {
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "Retry Task",
        "operationId": "retry_task_api_v1_tasks__task_id__retry_post",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskCreateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tasks/{task_id}/reports": {
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "Create Report",
        "operationId": "create_report_api_v1_tasks__task_id__reports_post",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tasks/{task_id}/mask-frames/{slice_index}": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "Get Mask Frame",
        "operationId": "get_mask_frame_api_v1_tasks__task_id__mask_frames__slice_index__get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          },
          {
            "name": "slice_index",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Slice Index"
            }
          },
          {
            "name": "prefix",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^[a-zA-Z0-9_-]{1,32}$",
              "default": "label",
              "title": "Prefix"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tasks/{task_id}/artifacts/{name}": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "Get Artifact",
        "operationId": "get_artifact_api_v1_tasks__task_id__artifacts__name__get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          },
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ArtifactRef": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "uri": {
            "type": "string",
            "title": "Uri"
          },
          "media_type": {
            "type": "string",
            "title": "Media Type",
            "default": "application/octet-stream"
          },
          "size_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Size Bytes"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          }
        },
        "type": "object",
        "required": [
          "name",
          "uri"
        ],
        "title": "ArtifactRef"
      },
...
      "ClassificationPrediction": {
        "properties": {
          "label": {
            "type": "string",
            "title": "Label"
          },
          "probability": {
            "type": "number",
            "title": "Probability"
          }
        },
        "type": "object",
        "required": [
          "label",
          "probability"
        ],
        "title": "ClassificationPrediction"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "InferenceResult": {
        "properties": {
          "type": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "segmentation",
                  "detection",
                  "classification"
                ]
              },
              {
                "type": "string"
              }
            ],
            "title": "Type"
          },
          "series_uid": {
            "type": "string",
            "title": "Series Uid"
          },
          "model_id": {
            "type": "string",
            "title": "Model Id"
          },
          "model_version": {
            "type": "string",
            "title": "Model Version"
          },
          "runtime_ms": {
            "type": "integer",
            "title": "Runtime Ms"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "artifacts": {
            "items": {
              "$ref": "#/components/schemas/ArtifactRef"
            },
            "type": "array",
            "title": "Artifacts"
          },
          "masks": {
            "items": {
              "$ref": "#/components/schemas/MaskArtifact"
            },
            "type": "array",
            "title": "Masks"
          },
          "boxes": {
            "items": {
              "$ref": "#/components/schemas/DetectionBox"
            },
            "type": "array",
            "title": "Boxes"
          },
          "predictions": {
            "items": {
              "$ref": "#/components/schemas/ClassificationPrediction"
            },
            "type": "array",
            "title": "Predictions"
          },
          "cam_overlay_uri": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cam Overlay Uri"
          }
        },
        "type": "object",
        "required": [
          "type",
          "series_uid",
          "model_id",
          "model_version",
          "runtime_ms",
          "summary"
        ],
        "title": "InferenceResult"
      },
...
      "Page_TaskSummary_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TaskSummary"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "Page[TaskSummary]"
      },
...
      "ReportCreateRequest": {
        "properties": {
          "finding_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Finding Ids"
          },
          "export_seg": {
            "type": "boolean",
            "title": "Export Seg",
            "default": true
          },
          "export_sr": {
            "type": "boolean",
            "title": "Export Sr",
            "default": true
          },
          "export_gsps": {
            "type": "boolean",
            "title": "Export Gsps",
            "default": true
          }
        },
        "type": "object",
        "title": "ReportCreateRequest",
        "description": "Generate structured report (+ optional DICOM exports) from selected findings."
      },
      "ReportResponse": {
        "properties": {
          "report_id": {
            "type": "string",
            "title": "Report Id"
          },
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "model_id": {
            "type": "string",
            "title": "Model Id"
          },
          "model_version": {
            "type": "string",
            "title": "Model Version"
          },
          "finding_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Finding Ids"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "artifacts": {
            "items": {
              "$ref": "#/components/schemas/ReportArtifact"
            },
            "type": "array",
            "title": "Artifacts"
          }
        },
        "type": "object",
        "required": [
          "report_id",
          "task_id",
          "model_id",
          "model_version",
          "text"
        ],
        "title": "ReportResponse"
      },
...
      "TaskCreateRequest": {
        "properties": {
          "series_uid": {
            "type": "string",
            "title": "Series Uid"
          },
          "model_id": {
            "type": "string",
            "title": "Model Id"
          },
          "params": {
            "additionalProperties": true,
            "type": "object",
            "title": "Params"
          }
        },
        "type": "object",
        "required": [
          "series_uid",
          "model_id"
        ],
        "title": "TaskCreateRequest"
      },
      "TaskCreateResponse": {
        "properties": {
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "cache_hit": {
            "type": "boolean",
            "title": "Cache Hit",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "task_id",
          "status"
        ],
        "title": "TaskCreateResponse"
      },
...
      "TaskSummary": {
        "properties": {
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "series_uid": {
            "type": "string",
            "title": "Series Uid"
          },
          "study_uid": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Study Uid"
          },
          "model_id": {
            "type": "string",
            "title": "Model Id"
          },
          "model_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Version"
          },
          "params": {
            "additionalProperties": true,
            "type": "object",
            "title": "Params"
          },
          "status": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "queued",
                  "running",
                  "succeeded",
                  "failed",
                  "canceled"
                ]
              },
              {
                "type": "string"
              }
            ],
            "title": "Status"
          },
          "stage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage"
          },
          "progress": {
            "type": "number",
            "title": "Progress",
            "default": 0.0
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "cache_hit": {
            "type": "boolean",
            "title": "Cache Hit",
            "default": false
          },
          "cached_from": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cached From"
          },
          "runtime_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Runtime Ms"
          },
          "stage_timings": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage Timings"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "artifacts": {
            "items": {
              "$ref": "#/components/schemas/ArtifactRef"
            },
            "type": "array",
            "title": "Artifacts"
          },
          "logs": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/TaskLog"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Logs"
          }
        },
        "type": "object",
        "required": [
          "task_id",
          "series_uid",
          "model_id",
          "status"
        ],
        "title": "TaskSummary"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}